--- a/src/classes/atom.cpp
+++ b/src/classes/atom.cpp
@@ -103,7 +103,7 @@ void Atom::translateCoordinates(double dx, double dy, double dz) { setCoordinate
  */
 
 // Return scaling factor to employ with specified Atom
-double Atom::scaling(const std::shared_ptr<Atom> &j) const
+double Atom::scaling(const Atom *j) const
 {
     assert(speciesAtom_ != nullptr);
     assert(j != nullptr);