--- a/src/classes/atom.h
+++ b/src/classes/atom.h
@@ -104,5 +104,5 @@ class Atom : public DynamicArrayObject<Atom>
      */
     public:
     // Return scaling factor to employ with specified Atom
-    double scaling(const std::shared_ptr<Atom> &j) const;
+    double scaling(const Atom *j) const;
 };