--- a/src/classes/configuration.h
+++ b/src/classes/configuration.h
@@ -187,7 +187,7 @@ class Configuration : public ListItem<Configuration>
     // Update Cell contents
     void updateCellContents();
     // Update Cell location of specified Atom
-    void updateCellLocation(const std::shared_ptr<Atom> &i);
+    void updateCellLocation(Atom *i);
     // Update Cell location of specified Molecule
     void updateCellLocation(const std::shared_ptr<Molecule> &mol);
     // Update Cell location of specified Atom indices