--- a/src/modules/energy/process.cpp
+++ b/src/modules/energy/process.cpp
@@ -118,7 +118,7 @@ bool EnergyModule::process(Dissolve &dissolve, ProcessPool &procPool)
                             continue;
 
                         // Get intramolecular scaling of atom pair
-                        scale = i->scaling(j);
+                        scale = i->scaling(j.get());
                         if (scale < 1.0e-3)
                             continue;