… benchmark/benchmark_files/hexane5k.txt