This PR performs some further (and much better) optimisation on the `EnergyKernel::energy(Molecule &mol, ...)` function, improving the performance of the core `MolShake` module.

Timing information is as follows, run on a 48-core machine. Speed increases vs the serial version are nowhere near 48x, but this is perhaps unsurprising given the localised nature of the routine making it difficult to scale.

```
------------------------------------------------------------------------------
Serial @ 9093bf7 (first optimisation)        Time             CPU   Iterations
------------------------------------------------------------------------------
smallMolecule, Population::small>        0.260 ms        0.260 ms         2491
mediumMolecule, Population::small>        2.12 ms         2.12 ms          335
mediumMolecule, Population::medium>       1.80 ms         1.80 ms          394

------------------------------------------------------------------------------
Threaded @ 9093bf7 (first optimisation)      Time             CPU   Iterations
------------------------------------------------------------------------------
smallMolecule, Population::small>        0.452 ms        0.452 ms         1485
mediumMolecule, Population::small>        2.82 ms         2.82 ms          226
mediumMolecule, Population::medium>       1.77 ms         1.77 ms          350

------------------------------------------------------------------------------
Serial @ Current                             Time             CPU   Iterations
------------------------------------------------------------------------------
smallMolecule, Population::small>        0.199 ms        0.199 ms         3533
mediumMolecule, Population::small>        1.10 ms         1.10 ms          638
mediumMolecule, Population::medium>      0.900 ms        0.900 ms          753

------------------------------------------------------------------------------
Threaded @ Current                           Time             CPU   Iterations
------------------------------------------------------------------------------
smallMolecule, Population::small>        0.047 ms        0.047 ms        11028
mediumMolecule, Population::small>       0.212 ms        0.212 ms         3090
mediumMolecule, Population::medium>      0.237 ms        0.237 ms         2766
```